Chez Scheme: add "reboot" bootstrap
A "reboot" bootstrap is like "rktboot", but based on an existing Chez
Scheme instead of an existing Racket. Like "rktboot", "reboot"
implements enough of Chez Scheme to run on a relatively unrelated Chez
Scheme implementation, as opposed to the normal "cross boot" process
that needs a Chez Scheme with the same record representation,
intermediate structures for expanded code, etc.