move `begin-unsafe` to `$begin-unsafe`
While the `begin-unsafe` form seems like it might be a useful idea,
the current implementation doesn't work in definition positions, and
it's not clear how well it fits with macro expansion in general. For
now, hide it as a system `$begin-unsafe` form to be completed or
replaced with a better approach in the future.