repair mutex release on error
When the kernel escapes with an error, it's meant to release all of
its acquites on the tc mutex, but changes in 3a9255b1a1 removed
a counter that distinguishes kernel acquires from Scheme acquires.

A test for this is part of a subsequent `gensym->unique-string`
commit, since `gensym->unique-string` takes a mutex around the call to
a kernel function that errors.