Chez Scheme: adjust memory use of building with pb
When building Chez Scheme with itself, the collect request handler is
configured to trade space for speed. The space trade-off turns out to
be up to a factor of 2. There's little speed improvement for pb,
though, since interpreter time dominates. Meanwhile, pb on a 32-bit
platform already doubles memory use by using 64-bit addresses. So,
disable the collect request specialization when running on pb, and
that accomodates 32-bit hosts that have less memory.