repair `char-foldcase` for some single-character special cases
For example, `(char-foldcase #\x1E9E)` should be `#\xDF`, in contrast
to the `(string-foldcase (string #\x1E9E))` result of "ss". The
character foldings that that are different here correspond to cases
that are "S" status in "CaseFolding.txt".

Thanks to Marc Feeley for noticing and reporting the bug.

Also, repair some dependencies in "s/build.zuo".