represent some flonums as immediates
Based on "Float Self-Tagging" by Melançon, Serrano, and Feeley.

A difference here is that the tag is not a "self" tag. Chez Scheme's
general tagging system is unchanged, so #b010 is still the tag for a
flonum. The distinction between an immediate and non-immediate flonum
is encoded by a fourth bit: #b1010 means a pointer as before, while

This additional tagging bit for flonums requires that flonums are
allocated on doubleword boundaries. Note that #b1010 is the tag for a
pointer because tags are formed by subtracting from an address, not
adding, and adding to a value that ends in #b1010 will arrive at a
doubleword-aligned address. Flonums were already doubleword-aligned,
except when extracted as the imaginary part of an inexactnum (i.e.,
inexact complex number); that special extraction shortcut is disabled.
On the plus side, keeping allocated flonums doubleword-aligned
simplifies forwarding checks in the GC.

The immediately represented flonums are the ones who have all the same
bits for the 5 next-to-high exponent bits in the IEEE representation,
marked here as "E":

 seEEEEEeeeeemmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm

This matches flonums whose magnitude is less than 1e-300 or so
(including 0 and all subnormals), whose magnitude is between 1e-9 and
1e9 or so, whose magnitude is above 1e300 or so, and the infinities
and NaNs.

The encoded form of such numbers is rotated left by 6 bits, then with
the low bits replaced by the tag.

 Eeeeeemmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmse0010

The high E encodes the dropped bits. An arithmetic shift right by 4
bits restores the dropped Es, and then a right rotation by 2 completes
decoding.

The implementation of encoding involves two branches to check whether
EEEEE is 00000, 11111, or something else. Those two branches were
faster than other strategies I tried, probably because they are
predictable branches in example programs.