add `ftype-object-pointer` and related foreign-pointer extensions
A "foreign object pointer" is a reference to a bytevector or flvector
that works like a foreign pointer and produces the address of the
vector content (like `object->reference-address`), but with GC
cooperation so that the address does not go stale --- as long as it is
extracted at the last moment in a context where a collection cannot
occur.

With object pointers as a way to unify GCable and truly foreign
references through the ftype interface, some further additions allow
the ftype layer to include more of the even less safe functionality of
the lower-level `foreign-ref` API, which extracts data from a
reference without a declared and checked foreign representaton on the
reference. The `ftype-any-ref` and `ftype-any-set!` forms cover
pointer access and update, and `(* void)` is allowed as a generic
pointer type description (e.g., in `foreign-procedure` and
foreign-callable`).