improve `string-append` for long argument lists
Internally, when more than two arguments are provided to
`string-append`, it used a non-tail call to sum the string sizes
recursively and then copy strings into place on return. This commit
changes `string-append` to use two iterative traversals, except for
the 2-string case that was already treated specially.

When the number of string is more than some number (6 on my current
machine), the allocation of continuation frames becomes more expensive
than a second iterative traversal of a list. For large number of
relatively small strings, two iterative traversals can be about twice
as fast as allocating a large continuation. For a small number of
strings (i.e., less than 6) or larger strings, two iterative
traversals is only slightly slower. Given that the 2-sting case is
handled separately, a small slowdown for a few strings seems a
worthwhile tradeoff to substantially speed up the case of many
strings.