add `__atomic` foreign convention and improve related flonum unboxing
The `__atomic` convention promises that a foreign function will not
call back into Scheme, such as through a foreign callable, before it
returns. That promise enables a more direct call to the foreign
function. In particular, it allows the call to be exposed enough in
intermediate compiler passes so that the call can partitcapte in
flonum-unboxing optimizations (when the foreign function has arguments
or results that are represented as flonums).

In principle, unboxing optimizations could also apply to
non-`__atomic` foreign functions, but the current calling convention
for Scheme functions and the protocol between Scheme code and the
kernel (especially the garbage collector) make that difficult. Further
improvements there are possible, but functions that fit the `__atomic`
constraint can be handled relatively easily.

[Temporary: will try changing back to `foreign-procedure` as the
primitive]
Internally, the `foreign-procedure` primitive form has been replaced
with a `foreign-call` form; the drawback is that `foreign-call` is a
little more work to handle in early compiler passes (roughly
duplicating some work performed for `call`), but the advantage is that
the argument and result types are more readily exposed.

Meanwhile, various problems interfered with flonum unboxing around
`ftype-ref` and `ftype-set!` on `double` and `float` fields, and those
are now fixed.